backend/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
bitarray==1.2.2
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cytoolz==0.11.2
Django==4.0.6
django-cors-headers==3.13.0
django-filter==22.1
django-rest-authtoken==2.1.4
djangorestframework==3.13.1
drf-yasg==1.20.0
eth-abi==2.1.1
eth-account==0.5.8
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
frozenlist==1.3.0
hexbytes==0.2.2
idna==3.3
importlib-metadata==4.12.0
inflection==0.5.1
ipfshttpclient==0.8.0a2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.6.2
lru-dict==1.1.7
Markdown==3.3.7
MarkupSafe==2.1.1
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
packaging==21.3
parsimonious==0.8.1
protobuf==3.20.1
pycryptodome==3.15.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
requests==2.28.1
rlp==2.0.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
toolz==0.11.2
uritemplate==4.1.1
urllib3==1.26.10
varint==1.0.2
web3==5.30.0
websockets==9.1
yarl==1.7.2
zipp==3.8.0
